Rename matrix_slice.rs to matrix_view.rs
